requirements.txt
aws-cdk-lib==2.99.1
constructs==10.1.162
PyYAML==6.0
cdk-nag==2.27.178